code tweaks
